+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 2.9.9
 ----------------
 - Fix PHP error when no options set for multiple field in Referral.php & MakeReferral.php
+- Fix SQL error when all-day checked & minutes in Calendar.php
 
 Changes in 2.9.8
 ----------------